Enrich persistence state!  v0.3.0
The application now serialises an AppState structure, which contains
enum information about different states of applications, as well as
the current mode the application is running in (currently binary).
I.e.:

Modes:
  - ToggleMode::Hide
  - ToggleMode::Show

Application states:
  - HiddenState::Hidden
  - HiddenState::NotHidden
  - HiddenState::AlreadyHidden

Enriching the application state as I have done in this commit allows
better information to be gleaned from the persistance file.  This
means we have better control over actions required for the app to
work (rather than trying to infer what toggle state we are in).

We also factor out unsafe system-access code into separate functions
and push the results into a struct, for easy access.

These changes make the app more easily extensible in the future,
as many things now exist in structs and enums.