Empty app state persistance on complete toggle!  v0.3.1
After one full cycle of the application (i.e., hide and then show
again), we don't want the app state interfering with the next run of
the application, so we need to empty it at this point.

Logically equivalently: we only write the app state if we are half way
through the toggle state.